Move FullyBakedTx and RawTx to sov-rollup-interface
Moves FullyBakedTx and RawTx from sov-modules-api to sov-rollup-interface/stf
to establish proper architectural layering. The rollup-interface crate is
the foundational layer for transaction types, and the stf module is where
state transition transaction types belong alongside TransactionReceipt.

Changes:
- Created crates/rollup-interface/src/state_machine/stf/batch.rs with transaction types
- Updated stf/mod.rs to export batch types
- Updated sov-modules-api to re-export from rollup-interface for backward compatibility
- Added module-level documentation

This is a non-breaking change that prepares for making TransactionReceipt.body_to_save
type-safe by storing FullyBakedTx directly instead of raw Bytes.
