+++ crates/module-system/sov-modules-api/src/batch.rs
@@ -8,86 +8,10 @@ use crate::{
 use borsh::{BorshDeserialize, BorshSerialize};
 use serde::{Deserialize, Serialize};
 use sov_rollup_interface::da::DaSpec;
-use sov_rollup_interface::Bytes;
 
-/// `FullyBakedTx` represents a serialized signed rollup transaction that has been encoded with
-/// authentication information and is ready to be placed on the DA layer.
-#[derive(
-    PartialEq,
-    Eq,
-    Clone,
-    BorshDeserialize,
-    BorshSerialize,
-    Serialize,
-    Deserialize,
-    derive_more::AsRef,
-)]
-pub struct FullyBakedTx {
-    /// Serialized transaction.
-    #[as_ref(forward)]
-    pub data: Bytes,
-    /// Sequencer-provided metadata for each transaction (e.g., timestamps).
-    /// This data is NOT signed by users but is added by the sequencer.
-    pub sequencing_data: Option<Bytes>,
-}
-
-impl std::fmt::Debug for FullyBakedTx {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("FullyBakedTx")
-            .field("data", &hex::encode(&self.data))
-            .finish()
-    }
-}
-
-impl FullyBakedTx {
-    /// Construct a `FullyBakedTx` containing the given data
-    #[must_use]
-    pub fn new(data: Vec<u8>) -> Self {
-        Self {
-            data: Bytes::from_owner(data),
-            sequencing_data: None,
-        }
-    }
-
-    /// Sets sequencing metadata
-    pub fn set_sequencing_metadata(&mut self, metadata: &impl BorshSerialize) {
-        self.sequencing_data = Some(borsh::to_vec(metadata).unwrap().into());
-    }
-}
-
-/// `RawTx` represents a serialized signed rollup transaction. A `RawTx` needs to be encoded
-/// with authentication information before being placed on the DA layer.
-#[derive(
-    PartialEq,
-    Eq,
-    Clone,
-    BorshDeserialize,
-    BorshSerialize,
-    Serialize,
-    Deserialize,
-    derive_more::AsRef,
-)]
-pub struct RawTx {
-    /// Serialized transaction.
-    #[as_ref(forward)]
-    pub data: Vec<u8>,
-}
-
-impl std::fmt::Debug for RawTx {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("RawTx")
-            .field("data", &hex::encode(&self.data))
-            .finish()
-    }
-}
-
-impl RawTx {
-    /// Construct a `RawTx` containing the given data
-    #[must_use]
-    pub fn new(data: Vec<u8>) -> Self {
-        Self { data }
-    }
-}
+// Re-export FullyBakedTx and RawTx from rollup-interface for backward compatibility
+pub use sov_rollup_interface::stf::{FullyBakedTx, RawTx};
+pub use sov_rollup_interface::Bytes;
 
 /// A blob that has been selected for execution
 pub struct SelectedBlob<S: Spec, B> {
